refactor: centralize LNURL response types
